Move BC Assessment tab
