+++ source/frontend/src/features/mapSideBar/property/PropertyContainer.tsx
@@ -1,5 +1,5 @@
 import { usePropertyDetails } from 'features/mapSideBar/hooks/usePropertyDetails';
-import BcAssessmentTabView from 'features/mapSideBar/tabs/bcAssessment/BcAssessmentTabView';
+import BcAssessmentTabView from 'features/mapSideBar/property/tabs/bcAssessment/BcAssessmentTabView';
 import {
   InventoryTabNames,
   InventoryTabs,
