+++ source/frontend/src/features/mapSideBar/property/tabs/bcAssessment/AssessedValuesTable.tsx
