+++ source/frontend/src/features/mapSideBar/property/tabs/bcAssessment/BcAssessmentTabView.test.tsx
