+++ source/frontend/src/features/mapSideBar/property/tabs/bcAssessment/SalesTable.tsx
